requirements.txt
aiohttp
aiohttp-jinja2
praw
pymongo
python-dateutil
python-twitter
